# backend/openapi.json
{"openapi":"3.0.2","info":{"title":"Segundo parcial. Ingeniería web","description":"Altair Bueno Calvente. Software 4ºA","version":"0.1.0"},"paths":{"/example":{"get":{"summary":"Example","operationId":"example_example_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/geocoding":{"get":{"summary":"Geocoding","operationId":"geocoding_geocoding_get","parameters":[{"required":true,"schema":{"title":"Query","type":"string"},"name":"query","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/requires_auth":{"get":{"summary":"Requires Auth","operationId":"requires_auth_requires_auth_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/orders":{"post":{"tags":["Orders"],"summary":"Create Order","description":"Creates a new order","operationId":"create_order","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrder"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/orders/{orderId}/capture":{"put":{"tags":["Orders"],"summary":"Capture","operationId":"capture","parameters":[{"required":true,"schema":{"title":"Orderid","type":"string"},"name":"orderId","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaptureOrder"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}}},"components":{"schemas":{"CaptureOrder":{"title":"CaptureOrder","type":"object","properties":{}},"CreateOrder":{"title":"CreateOrder","type":"object","properties":{}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}},"servers":[{"url":"/","description":"Default"},{"url":"http://localhost:8000","description":"Docker compose"}]}